second commit

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require("../models/Thing");
+const fs = require("fs");
 /*
 Ici, vous créez une instance de votre modèle Thing en lui passant un objet JavaScript contenant toutes les informations requises du corps de requête 
 analysé (en ayant supprimé en amont le faux_id envoyé par le front-end).
@@ -77,9 +78,26 @@ La méthode deleteOne() de notre modèle fonctionne comme findOne() et updateOne
 Nous envoyons ensuite une réponse de réussite ou d'échec au front-end.
 */
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-        .catch((error) => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+        .then((thing) => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({ message: "Not authorized" });
+            } else {
+                const filename = thing.imageUrl.split("/images/")[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({ _id: req.params.id })
+                        .then(() => {
+                            res.status(200).json({
+                                message: "Objet supprimé !",
+                            });
+                        })
+                        .catch((error) => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch((error) => {
+            res.status(500).json({ error });
+        });
 };
 /*
 si aucun Thing n'est trouvé ou si une erreur se produit, nous envoyons une erreur 404 au front-end, avec l'erreur générée.
